dockers/IoT/requirements.txt
aiohttp
requests
pymongo
pyyaml
pytz
yarl==1.7.2